requirements.txt
# python-telegram-bot API install 
python-telegram-bot==7.0.1
requests==2.16.5